+++ src/main/java/com/hubo/gillajabi/crawl/infrastructure/persistence/CourseDetailRepository.java
@@ -1,0 +1,7 @@
+package com.hubo.gillajabi.crawl.infrastructure.persistence;
+
+import com.hubo.gillajabi.crawl.domain.entity.CourseDetail;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CourseDetailRepository  extends JpaRepository<CourseDetail, Long> {
+}
